requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
cattrs==24.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
flatbuffers==25.2.10
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==6.0.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.2.4
openmeteo_requests==1.3.0
openmeteo_sdk==1.19.0
openrouteservice==2.3.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-geohash==0.8.5
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
url-normalize==1.4.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
